requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
paho-mqtt==1.6.1
requests